WIP: Split up and organize scripts
