+++ src/consts.js
@@ -4,3 +4,11 @@ export const STYLE_CLASS = "fbtrStyled";
 export const PROCESSED_CLASS = "FBTR-PROCESSED";
 
 export const NOOP = () => { };
+
+export const MSG = Object.freeze({
+    chromeShow: "CHROME_SHOW",
+    history: "HISTORY",
+    insertCss: "INSERT_CSS",
+    removeCss: "REMOVE_CSS",
+    rulesRefresh: "RULES",
+});
